Introduce userController, do user specific actions in there
